Travis build configured to automatically publish to Github pages.
Note that this only currently publishes on success. You can add an after_failure section to .travis.yml to also publish on failure.
Also, this publishes the spec folder to github, rather than the sub-folder which publish.gradle currently publishes. This makes it consistent with the other extensions, and wil mean that any additional specs are always picked up.
After publishing, the location of the spec in the README.md and on concordion-website will need to be updated to reference the longer path:
spec/org/concordion/ext/excel/tutorial